+++ vite.config.ts
@@ -1,27 +1,25 @@
+import { defineConfig } from "vite";
 import dts from "vite-plugin-dts";
 import { visualizer } from "rollup-plugin-visualizer";
-import { defineConfig } from "vite";
 
 export default defineConfig({
   plugins: [
-    dts({
-      rollupTypes: true,
-    }),
+    dts(),
   ],
   build: {
-    lib: {
-      fileName: "index",
-      entry: "src/index.ts",
-      formats: ["es", "cjs"],
-    },
-    sourcemap: true,
-    minify: true,
     rollupOptions: {
       plugins: [
         visualizer({
-          filename: "./build-report.local.html",
+          filename: "./dist/build-report.html",
         }),
       ],
     },
+    lib: {
+      fileName: "index",
+      entry: "./src/index.ts",
+      formats: ["es", "cjs"],
+    },
+    minify: true,
+    sourcemap: true,
   },
 });