+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # Changelog
 
-## 0.7.0
+## [0.8.0]
 
-**Features**
+### Added
+
+- `StonApiClient.getSwapPairs()` method
+- `StonApiClient.simulateSwap()` method
+- `StonApiClient.simulateReverseSwap()` method
+- `StonApiClient.getSwapStatus()` method
+
+### Deprecated
+
+- `baseUrl` client option deprecated. Expected `baseURL` to be used instead
+
+## [0.7.0]
+
+### Added
 
 - `StonApiClient.getAssets()` method
 - `StonApiClient.getFarms()` method